message

+++ src/main/java/com/example/InsurancePremiumCalculator/service/PolicyService.java
@@ -2,7 +2,9 @@ package com.example.InsurancePremiumCalculator.service;
 
 import com.example.InsurancePremiumCalculator.domain.Policy;
 
+import java.math.BigDecimal;
+
 public interface PolicyService {
 
-     double calculatePremium(Policy policy);
+     BigDecimal calculatePremium(Policy policy);
 }
